# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "KLIK  DIGITAL SINERGI API",
    "description": "Backend REST API for technical test in klik digital sinergi. Build with Nodejs + Express, MongoDB, ElasticSearch & integrated with midtrans payment gateway",
    "contact": {}
  },
  "host": "localhost:1000",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/payments": {
      "post": {
        "description": "(Private Route | Need JWT Token) create a new payment & integrated with midtrans payment gateway.",
        "summary": "create payment",
        "tags": [
          "payment"
        ],
        "operationId": "createpayment",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createpaymentrequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/createpayment"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "data": {
                  "_id": "5f6fffa51421390b247f1a0c",
                  "user": {
                    "_id": "5f6ffad9e13f7a22ec61c6d4",
                    "username": "faeshal",
                    "email": "faeshal@gmail.com"
                  },
                  "productName": "pecel lele",
                  "quantity": 2,
                  "price": 20000,
                  "total": 40000,
                  "createdAt": "2020-09-27T02:57:41.982Z",
                  "__v": 0
                },
                "chargeResponse": {
                  "status_code": "201",
                  "status_message": "GO-PAY transaction is created",
                  "transaction_id": "93519dea-24d2-4d1e-b95b-b82b23786529",
                  "order_id": "5f6fffa51421390b247f1a0c",
                  "merchant_id": "G998502637",
                  "gross_amount": "40000.00",
                  "currency": "IDR",
                  "payment_type": "gopay",
                  "transaction_time": "2020-09-27 09:57:41",
                  "transaction_status": "pending",
                  "fraud_status": "accept",
                  "actions": [
                    {
                      "name": "generate-qr-code",
                      "method": "GET",
                      "url": "https://api.sandbox.veritrans.co.id/v2/gopay/93519dea-24d2-4d1e-b95b-b82b23786529/qr-code"
                    },
                    {
                      "name": "deeplink-redirect",
                      "method": "GET",
                      "url": "https://simulator.sandbox.midtrans.com/gopay/ui/checkout?referenceid=l1XOms5Iaj&callback_url=http%3A%2F%2Flocalhost%3A1000%2F%3Forder_id%3D5f6fffa51421390b247f1a0c"
                    },
                    {
                      "name": "get-status",
                      "method": "GET",
                      "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/status"
                    },
                    {
                      "name": "cancel",
                      "method": "POST",
                      "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/cancel"
                    }
                  ]
                }
              }
            },
            "headers": {
              "Content-Security-Policy": {
                "type": "string",
                "default": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
              },
              "X-DNS-Prefetch-Control": {
                "type": "string",
                "default": "off"
              },
              "Expect-CT": {
                "type": "string",
                "default": "max-age=0"
              },
              "X-Frame-Options": {
                "type": "string",
                "default": "SAMEORIGIN"
              },
              "Strict-Transport-Security": {
                "type": "string",
                "default": "max-age=15552000; includeSubDomains"
              },
              "X-Download-Options": {
                "type": "string",
                "default": "noopen"
              },
              "X-Content-Type-Options": {
                "type": "string",
                "default": "nosniff"
              },
              "X-Permitted-Cross-Domain-Policies": {
                "type": "string",
                "default": "none"
              },
              "Referrer-Policy": {
                "type": "string",
                "default": "no-referrer"
              },
              "X-XSS-Protection": {
                "type": "string",
                "default": "0"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "1257"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"4e9-bRgLTXquegJxMXrjdpuYillZdCE\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 27 Sep 2020 02:57:42 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              }
            }
          }
        }
      },
      "get": {
        "description": "(Private Route | Need JWT Token) get all payment data from all user. ",
        "summary": "list payments",
        "tags": [
          "payment"
        ],
        "operationId": "listpayments",
        "deprecated": false,
        "produces": [
          "application/json; charset=utf-8"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/listpayments"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "success": true,
                "total": 2,
                "data": [
                  {
                    "_id": "5f6fffa51421390b247f1a0c",
                    "user": {
                      "_id": "5f6ffad9e13f7a22ec61c6d4",
                      "username": "faeshal",
                      "email": "faeshal@gmail.com"
                    },
                    "productName": "pecel lele",
                    "quantity": 2,
                    "price": 20000,
                    "total": 40000,
                    "createdAt": "2020-09-27T02:57:41.982Z",
                    "__v": 0
                  },
                  {
                    "_id": "5f7000701421390b247f1a0e",
                    "user": {
                      "_id": "5f70003c1421390b247f1a0d",
                      "username": "abdullah",
                      "email": "abdullah@gmail.com"
                    },
                    "productName": "martabak manis",
                    "quantity": 4,
                    "price": 10000,
                    "total": 40000,
                    "createdAt": "2020-09-27T03:01:04.626Z",
                    "__v": 0
                  }
                ]
              }
            },
            "headers": {
              "Content-Security-Policy": {
                "type": "string",
                "default": "default-src 'self';base-uri 'self';block-all-mixed-content;font-src 'self' https: data:;frame-ancestors 'self';img-src 'self' data:;object-src 'none';script-src 'self';script-src-attr 'none';style-src 'self' https: 'unsafe-inline';upgrade-insecure-requests"
              },
              "X-DNS-Prefetch-Control": {
                "type": "string",
                "default": "off"
              },
              "Expect-CT": {
                "type": "string",
                "default": "max-age=0"
              },
              "X-Frame-Options": {
                "type": "string",
                "default": "SAMEORIGIN"
              },
              "Strict-Transport-Security": {
                "type": "string",
                "default": "max-age=15552000; includeSubDomains"
              },
              "X-Download-Options": {
                "type": "string",
                "default": "noopen"
              },
              "X-Content-Type-Options": {
                "type": "string",
                "default": "nosniff"
              },
              "X-Permitted-Cross-Domain-Policies": {
                "type": "string",
                "default": "none"
              },
              "Referrer-Policy": {
                "type": "string",
                "default": "no-referrer"
              },
              "X-XSS-Protection": {
                "type": "string",
                "default": "0"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "523"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"20b-Z9+MXxIr3PBFyocTeo8Jiq8zn+Y\""
              },
              "Date": {
                "type": "string",
                "default": "Sun, 27 Sep 2020 03:01:41 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "createpaymentrequest": {
      "title": "createpaymentrequest",
      "example": {
        "productName": "pecel lele",
        "quantity": 2,
        "price": 20000
      },
      "type": "object",
      "properties": {
        "productName": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "productName",
        "quantity",
        "price"
      ]
    },
    "createpayment": {
      "title": "createpayment",
      "example": {
        "success": true,
        "data": {
          "_id": "5f6fffa51421390b247f1a0c",
          "user": {
            "_id": "5f6ffad9e13f7a22ec61c6d4",
            "username": "faeshal",
            "email": "faeshal@gmail.com"
          },
          "productName": "pecel lele",
          "quantity": 2,
          "price": 20000,
          "total": 40000,
          "createdAt": "2020-09-27T02:57:41.982Z",
          "__v": 0
        },
        "chargeResponse": {
          "status_code": "201",
          "status_message": "GO-PAY transaction is created",
          "transaction_id": "93519dea-24d2-4d1e-b95b-b82b23786529",
          "order_id": "5f6fffa51421390b247f1a0c",
          "merchant_id": "G998502637",
          "gross_amount": "40000.00",
          "currency": "IDR",
          "payment_type": "gopay",
          "transaction_time": "2020-09-27 09:57:41",
          "transaction_status": "pending",
          "fraud_status": "accept",
          "actions": [
            {
              "name": "generate-qr-code",
              "method": "GET",
              "url": "https://api.sandbox.veritrans.co.id/v2/gopay/93519dea-24d2-4d1e-b95b-b82b23786529/qr-code"
            },
            {
              "name": "deeplink-redirect",
              "method": "GET",
              "url": "https://simulator.sandbox.midtrans.com/gopay/ui/checkout?referenceid=l1XOms5Iaj&callback_url=http%3A%2F%2Flocalhost%3A1000%2F%3Forder_id%3D5f6fffa51421390b247f1a0c"
            },
            {
              "name": "get-status",
              "method": "GET",
              "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/status"
            },
            {
              "name": "cancel",
              "method": "POST",
              "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/cancel"
            }
          ]
        }
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "data": {
          "$ref": "#/definitions/Data2"
        },
        "chargeResponse": {
          "$ref": "#/definitions/ChargeResponse"
        }
      },
      "required": [
        "success",
        "data",
        "chargeResponse"
      ]
    },
    "Data2": {
      "title": "Data2",
      "example": {
        "_id": "5f6fffa51421390b247f1a0c",
        "user": {
          "_id": "5f6ffad9e13f7a22ec61c6d4",
          "username": "faeshal",
          "email": "faeshal@gmail.com"
        },
        "productName": "pecel lele",
        "quantity": 2,
        "price": 20000,
        "total": 40000,
        "createdAt": "2020-09-27T02:57:41.982Z",
        "__v": 0
      },
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "productName": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "createdAt": {
          "type": "string"
        },
        "__v": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "_id",
        "user",
        "productName",
        "quantity",
        "price",
        "total",
        "createdAt",
        "__v"
      ]
    },
    "User": {
      "title": "User",
      "example": {
        "_id": "5f6ffad9e13f7a22ec61c6d4",
        "username": "faeshal",
        "email": "faeshal@gmail.com"
      },
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "required": [
        "_id",
        "username",
        "email"
      ]
    },
    "ChargeResponse": {
      "title": "ChargeResponse",
      "example": {
        "status_code": "201",
        "status_message": "GO-PAY transaction is created",
        "transaction_id": "93519dea-24d2-4d1e-b95b-b82b23786529",
        "order_id": "5f6fffa51421390b247f1a0c",
        "merchant_id": "G998502637",
        "gross_amount": "40000.00",
        "currency": "IDR",
        "payment_type": "gopay",
        "transaction_time": "2020-09-27 09:57:41",
        "transaction_status": "pending",
        "fraud_status": "accept",
        "actions": [
          {
            "name": "generate-qr-code",
            "method": "GET",
            "url": "https://api.sandbox.veritrans.co.id/v2/gopay/93519dea-24d2-4d1e-b95b-b82b23786529/qr-code"
          },
          {
            "name": "deeplink-redirect",
            "method": "GET",
            "url": "https://simulator.sandbox.midtrans.com/gopay/ui/checkout?referenceid=l1XOms5Iaj&callback_url=http%3A%2F%2Flocalhost%3A1000%2F%3Forder_id%3D5f6fffa51421390b247f1a0c"
          },
          {
            "name": "get-status",
            "method": "GET",
            "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/status"
          },
          {
            "name": "cancel",
            "method": "POST",
            "url": "https://api.sandbox.veritrans.co.id/v2/93519dea-24d2-4d1e-b95b-b82b23786529/cancel"
          }
        ]
      },
      "type": "object",
      "properties": {
        "status_code": {
          "type": "string"
        },
        "status_message": {
          "type": "string"
        },
        "transaction_id": {
          "type": "string"
        },
        "order_id": {
          "type": "string"
        },
        "merchant_id": {
          "type": "string"
        },
        "gross_amount": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "payment_type": {
          "type": "string"
        },
        "transaction_time": {
          "type": "string"
        },
        "transaction_status": {
          "type": "string"
        },
        "fraud_status": {
          "type": "string"
        },
        "actions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Action"
          }
        }
      },
      "required": [
        "status_code",
        "status_message",
        "transaction_id",
        "order_id",
        "merchant_id",
        "gross_amount",
        "currency",
        "payment_type",
        "transaction_time",
        "transaction_status",
        "fraud_status",
        "actions"
      ]
    },
    "Action": {
      "title": "Action",
      "example": {
        "name": "generate-qr-code",
        "method": "GET",
        "url": "https://api.sandbox.veritrans.co.id/v2/gopay/93519dea-24d2-4d1e-b95b-b82b23786529/qr-code"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "method": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "method",
        "url"
      ]
    },
    "listpayments": {
      "title": "listpayments",
      "example": {
        "success": true,
        "total": 2,
        "data": [
          {
            "_id": "5f6fffa51421390b247f1a0c",
            "user": {
              "_id": "5f6ffad9e13f7a22ec61c6d4",
              "username": "faeshal",
              "email": "faeshal@gmail.com"
            },
            "productName": "pecel lele",
            "quantity": 2,
            "price": 20000,
            "total": 40000,
            "createdAt": "2020-09-27T02:57:41.982Z",
            "__v": 0
          },
          {
            "_id": "5f7000701421390b247f1a0e",
            "user": {
              "_id": "5f70003c1421390b247f1a0d",
              "username": "abdullah",
              "email": "abdullah@gmail.com"
            },
            "productName": "martabak manis",
            "quantity": 4,
            "price": 10000,
            "total": 40000,
            "createdAt": "2020-09-27T03:01:04.626Z",
            "__v": 0
          }
        ]
      },
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Data2"
          }
        }
      },
      "required": [
        "success",
        "total",
        "data"
      ]
    }
  },
...
}